refactor: extract the library work stream and reframe the demo backend
Per the check-in: the SDK and the demo are separate work streams, and the
word SDK is retired inside the demo.

- sdk/ workspace removed; its six load-bearing modules re-homed as the
  private demo-backend/ workspace (privateState, profileProtocol, passkey,
  injection, encoding, types) with the PWA importing through a single seam
  (examples/passport-demo/src/backend.ts).
- signet adapter, its tests, the four library-stream documents, and the
  SDK explainer page leave with the library — preserved verbatim on the
  sdk/passport-sdk-handover branch.
- docs/sdk/ demo content moved to docs/demo/ and reframed; /sdk route,
  manifest shortcut, and SDK-named UI removed.
- WHAT-THIS-IS.md added: preview/testnet only, not mainnet, not the final
  product, not audited, and real rather than mocked (draft, pending
  ratification).

Gates: npm install, test:demo-backend (10/10), test:c1, test:pwa (40
checks), build, and the no-SDK greps all pass from a clean install.
