+++ WHAT-THIS-IS.md
@@ -1,0 +1,15 @@
+**What this is.** The Midnight Passport demo is a working prototype running on
+preview/testnet. It is real — the flows execute against real infrastructure, nothing is
+mocked. It exists to show that this can be done, and to show the functionality we are
+building towards.
+
+**What this is not.** It is not going to mainnet. It is not the final product. It has not
+been audited: audits and security hardening happen before anything moves onto a production
+path. Productising Passport is a separate stream of work running in parallel with this
+demo, not a phase that starts when the demo ends.
+
+**What we call things.** The demo engine is a demo backend with connectors. Integrations
+with partner applications are *connectors*, built case by case. Nothing in this demo is an
+SDK.
+
+Status: draft, 2026/07/29 — wording pending ratification (proposal: Karmel).
