+++ examples/passport-demo/package.json
@@ -25,7 +25,6 @@
     "@midnight-ntwrk/midnight-js-indexer-public-data-provider": "4.0.4",
     "@midnight-ntwrk/midnight-js-network-id": "4.0.4",
     "@midnight-ntwrk/midnight-js-types": "4.0.4",
-    "@midnight-ntwrk/passport-sdk": "file:../../sdk",
     "@midnight-ntwrk/wallet-sdk-address-format": "3.1.1",
     "@midnight-ntwrk/wallet-sdk-dust-wallet": "4.0.0",
     "@midnight-ntwrk/wallet-sdk-facade": "4.0.0",
@@ -37,6 +36,7 @@
     "bip39": "^3.1.0",
     "buffer": "^6.0.3",
     "lucide-react": "^0.468.0",
+    "passport-demo-backend": "file:../../demo-backend",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "rxjs": "^7.8.2",
