+++ package.json
@@ -4,14 +4,14 @@
   "private": true,
   "type": "module",
   "workspaces": [
-    "sdk",
+    "demo-backend",
     "examples/passport-demo",
     "examples/passport-profile-client"
   ],
   "scripts": {
-    "build": "npm run build --workspace @midnight-ntwrk/passport-sdk && npm run build --workspace passport-demo && npm run build --workspace passport-profile-client",
-    "test": "npm run test --workspace @midnight-ntwrk/passport-sdk",
-    "test:sdk": "npm run test --workspace @midnight-ntwrk/passport-sdk",
+    "build": "npm run build --workspace passport-demo-backend && npm run build --workspace passport-demo && npm run build --workspace passport-profile-client",
+    "test": "npm run test --workspace passport-demo-backend",
+    "test:demo-backend": "npm run test --workspace passport-demo-backend",
     "test:c1": "npm run test:c1 --workspace passport-demo",
     "test:pwa": "npm run test:pwa --workspace passport-demo",
     "demo": "npm run dev --workspace passport-demo",
